fix: adding structure
